# Speeded up json serialization on setting maintenance method