Serialization takes about 70% od setMaintenance function according to profiling. This PR replaces standard encoding json lib with [easyjson](https://github.com/mailru/easyjson) ([benchmark](https://yalantis.com/blog/speed-up-json-encoding-decoding/)) in order to improve performance for that function.

Here is a performance acceleration based on Github action tests measurements (smaller is better):
Before (with standard encoding/json):
BenchmarkSetTriggerCheckMaintenance-2   	    2062	    633651 ns/op
BenchmarkSetTriggerCheckMaintenance-2   	    1954	    634803 ns/op

After (with easyjson):
BenchmarkSetTriggerCheckMaintenance-2   	    3375	    348615 ns/op
BenchmarkSetTriggerCheckMaintenance-2   	    4041	    298578 ns/op