@dmitryanchikov 2022-04-29
Closing this as [another PR](https://github.com/moira-alert/moira/pull/745) is more relevant for the problem